Fix inaccuracies in subscription show response
This changes the subscription show response as it was inaccurate and
meant tests could pass or code could work (d'oh).

This response is still not fully valid as lots more fields are returned
that this, however changing all of this seems inconsistent with other
stubbed responses here which are quite minimal.

It includes additional options so that tests can represent more
scenarios.