Enhance the SpecialRoutePublisher to check first
Previously it would always create a new draft (or update the existing
one), and then publish it.

Now it checks if the content or links have changed, and only performs
the appropriate update actions if this is the case.

This code is often invoked when deploying applications, and most of
the time it runs, it doesn't actually change the content or links in a
meaningful way. However, it does create a new needless edition, and
this does lead to unnecessary Publishing API activity (due to
dependency resolution/link expansion). This is especialy true for the
homepage.