Add generate argument to `has_expanded_links` test helper
Now `publishing_api_has_expanded_links` can be called with `generate` as
an argument. This relates to the changes made in this pr for whitehall:
https://github.com/alphagov/whitehall/pull/4094.

`generate` defaults to `false` and in the real request would generate the
expanded links on the fly instead of getting them from stored data if set
to `true`.