Avoid using 'Link' header to unroll API pagination
The 'Link' header originally exposed by Whitehall is no longer reliable
- we think this is due to a Fastly issue. Currently this header is used
to compact a set of paginated API responses into a single array result,
using the "next" and "previous" links contained within the header.

Pagination is an issue for the /organisations and /world-locations APIs,
both of which also have pagination fields in the body, as well as in the
special 'Link' header. Using the 'next_page_url' and 'previous_page_url'
fields to control the compaction process fixes the 'Link' header bug.

In trying to make this change, the API helpers for organisations and world
locations needed updating to include the additional fields. In order to
make this change easier, I've refactored the code by combining similar
inline conditionals into discrete blocks of code. The 'Link' header is
no longer in the stub response, to reflect that it's not reliable.