Create GdsApi factory methods
This creates helper methods for the GdsApi module for initialising
adapters with common options.

The reason for this is that we have lots of the same boilerplate code in
every GOV.UK application to achieve the basic initialisations in this
file. This feels that it's more the responsibility of the gem if every
client has the same code.

Rather than an application needing to have:

```
require 'gds_api/publishing_api_v2'

GdsApi::PublishingApiV2.new(
  Plek.new.find("publishing-api"),
  bearer_token: ENV["PUBLISHING_API_BEARER_TOKEN"] || "example"
)
```

They can instead do `GdsApi.publishing_api_v2`