Change how the default logger is assigned
Previously, using `self.class.logger` meant that changes to
`GdsApi::Base.logger` wouldn't always apply to new clients.

I'm not to sure why, as Ruby is hard, but I am sure that this
behaviour is confusing. I also think it's unnecessary, as subclasses
can override the logger by passing it in as an option.

So, change this to use `GdsApi::Base.logger`, like how the
`default_options` works.