Change the expected Publishing API behaviour around included URLs
Updating Rails from 5.1 to 5.2 for the Publishing API seems to have
changed the URLs which are included in responses, for example, for the
next page of results.

I've been trying to decide if this is better or worse, and I think
that it's better. Mostly because percent encoding normally occurs when
a URL is created from it's component parts [1], and these are full URLs.

1: https://tools.ietf.org/html/rfc3986#page-14