Add bulk Asset Manager stubs
This adds in methods to do bulk actions to Asset Manager, the previous
mocks only support a single item at a time. This can be tricky in an
application such as Content Publisher where we may well be making
multiple calls to Asset Manager everytime we do anything (e.g. we send
the same image in 3 different sizes).

This ports in methods that were wrote explicitly in a helper for Content
Publisher.