Rename asset_manager_is_down
This changes the method name to be stub_asset_manager_isnt_available
which is consistent with a lot of the other test helpers, whereas the is
down was brand new terminology - oops.

This also removes the alias for the is_down method since it never
existed in a non stub_* form.