Add republish method for publishing api
This change adds the ability to use the republish enpointi in the
publishing api which is being introduced in the pr below. Calling this
method will send a request to pub api to republish a live document,
remove a withdrawal or send the published items data downstream.

Related to this pr: https://github.com/alphagov/publishing-api/pull/1439
Trello: https://trello.com/c/yhcKNEWr/587-allow-users-undo-withdraw