Rename subscribable to subscriber list
At one point we were planning to rename all instances of
subscriber lists with subscribables however these never happened
and it's confusing to have the same thing named differently.
Instead we've decided to stick with subscriber list.

This is a breaking change for users of the Email Alert API
adapters, clients would be expected to update the code. However,
Email Alert API itself will still accept the old style arguments.