Prevent warnings whenever gem is used
As the root file requires gds_api/rummager a warning in that class means
that every application that uses this gem receives a warning. This
changes this so that the warning only occurs when you initialise the
class. I removed the warning from the GdsApi.rummager method since you'd
receive one anyway by the call to new.

This also moves the warning for the test helper to only occur when this
module is included rather than any time the file is required. This
allows requiring the file without a warning but mixing this into
something else will trigger a warning.