Retry search requests on 5XX error
This will retry search requests a few times (with exponential
backoff) if the search query responds with a 500.

The motivation for this is that we see a fair number of
GdsApi::HTTPInternalServerError errors in our logs.
(Ive seen these are happening around deploys, but they
shouldnt ever happen, as they return the error to the user
and tells them to repeat the request).