Fix 'stub_email_alert_api_has_subscription' to stub /latest
The API added a '/latest' method here:
	https://github.com/alphagov/gds-api-adapters/pull/928

...but did not provide a way of stubbing it.
Instinctively it makes sense to add 'stub_email_alert_api_has_latest_subscription'
but that would mean writing our tests like this:

  stub_email_alert_api_has_subscription(1, 'weekly', :ended)
  stub_email_alert_api_has_latest_subscription(1, 'weekly', :ended)
  stub_email_alert_api_has_subscription(2, 'weekly')
  stub_email_alert_api_has_latest_subscription(2, 'weekly')

I think this makes things less readable - it can be misinterpreted
as "this subscription should be the latest one", which is not what
we want. What we want, if anything, is "this subscription should be
queryable against the '/latest' endpoint". But if you've stubbed the
email API for a subscription, that should just be done for you IMO.

_Every_ subscription should have a valid response if you call '/latest'.