Adds new 'stub_email_alert_api_has_subscriptions' method
We added a `get_latest_matching_subscription` method in https://github.com/alphagov/gds-api-adapters/pull/928 (v59.3.0), which takes a subscription ID and returns the latest matching subscription response (e.g. a user changing frequency from Daily to Weekly, if we query '/latest' on the Daily subscription, it will return the Weekly one as that is the most recent).

Consequently, the return value of '/latest' depends upon whether or not a newer subscription has been stubbed. It was very difficult / hacky / impossible to test the '/latest' method behaviour with the `get_latest_matching_subscription` method alone, so I've added a `stub_email_alert_api_has_subscriptions` which takes an array of subscriptions and stubs them in the order they are added.

I.e. if you pass `stub_email_alert_api_has_subscriptions([subscription1, subscription2])` and then query '/latest' on `subscription1`, the response for `subscription2` should be returned.