Forward X-Govuk-Authenticated-User-Organisation header
This middleware is added to allow
X-Govuk-Authenticated-User-Organisation to be passed to the
content-store when a user try's to access the draft gov.uk stack.

This header is passed from the authenticating-proxy after a user is
authenticated against signon, it contains the users organisation
content-id. This is needed as content-publisher is changing the way
access limiting of a document works by storing the access limited
organisation on the piece of content instead of storing all the users
uid's for an organisation on the content, although this will be still
possible.

Trello:
https://trello.com/c/wxp1XkSg/969-update-govuk-to-support-access-limiting-by-organisation