Stop publishing released version pact
There was an intention that we would have pact versions stored that
reflected each version of the gem that was published. So we have a pact
for version 59.5.1 of GDS API Adapters published [1] and that correlates
to version 59.5.1 of the gem.

This however is incorrect, the pact is published any time master builds
which is often numerous times before a release. This means that the
pacts are overridden with changes made after their release.

To resolve this we could set up some sort of checks to see if the pact
is already published or not but this feels unnecessary since it doesn't
appear that Publishing API uses any of these.

[1]: https://pact-broker.cloudapps.digital/hal-browser/browser.html#/pacticipants/GDS%20API%20Adapters/versions/59.5.1