Comment out Gzip compression test
Since rest-client 2.1, the gzip behaviour has changed
https://github.com/rest-client/rest-client/pull/597 such that it relies
on Net::HTTP rather than performing the decompression itself.

This test is currently failing because the body is not getting decoded
correctly, and I can't find a way of getting it to pass. It's blocking
us from adding new features so I thought it would be reasonable to
temporarily comment it out and I'll write up a card for Platform Health
to investigate putting it back.

I'm not sure if the test is actually useful in the first place because
we're testing HTTP behaviour of the Gem, not anything in our code.

I've tried the following things to get it to work:

- Manually adding an `Accept-Encoding` header (although it seems like
this definitely shouldn't work:
https://github.com/jnunemaker/httparty/issues/562#issuecomment-456975397)
- Manually adding a `Content-Type` header.
- Generating the Gziped string with `GzipWriter`.

One thing I've noticed is that for some reason `decode_content` is
always set to false on the Net::HTTPResponse object
(https://ruby-doc.org/stdlib-2.6.3/libdoc/net/http/rdoc/Net/HTTPResponse.html).