Correct response data on stub_any_publishing_api_path_reservation
This was returning data that was not what the Publishing API returns.
The Publishing API takes the input payload and merges in the base_path
(which is pretty weird/dubious for an API but that is not today's
problem): https://github.com/alphagov/publishing-api/blob/c1ee3b94a3abfa79b5800d22028c0129bb49ad64/app/commands/reserve_path.rb#L9

This also adds a test - which revealed a bug that is now fixed.