Fix responses and add test for put path validation failed stub
This corrects the responses of another of the path stubs for Publishing
API. It also adds in test and a docblock.

The somewhat funky string manipulation is done to replicate the
Publishing API response that is generated automatically by Rails. As
this gem doesn't depend on ActiveSupport we can't use the humanize
method.