Add a module for mixing in a deprecated alias
This offers an alternative to alias_method for when we have a deprecated
method. It will still allow the old method to be called but whenever it
is a warning will be emitted. This is intended to help us actually
remove some of the deprecated methods by encouraging their discontinued
use.

There is a reasonable chance this may have some problems with the switch
to Ruby 2.7 [1]. Hopefully we're well on the way to removing these
before then.

[1] https://eregon.me/blog/2019/11/10/the-delegation-challenge-of-ruby27.html