Switch the Worldwide requests to go to whitehall-frontend
The whitehall-frontend and whitehall-admin hosts should provide
identical responses, as they're the same code. The whitehall-frontend
hosts are setup for handling frontend requests, and have a read-only
database connection. The whitehall-admin hosts are setup for handling
the admin interface.

As the Worldwide API doesn't do any writing, it's suitable to be
handled by the whitehall-frontend hosts. This change is motivated by
the migration to AWS, as it'll be easier to keep whitehall-frontend up
during the migration (compared to whitehall-admin).