Fix mocha deprecation warning
In mocha-1.9 an undocumented feature was introduced where passing a
symbol as the first argument would create an unnamed mock with the
method represented by the symbol stubbed out. In 1.10 this behaviour
was reverted but it generates a deprecation warning which can be turned
off by setting `reinstate_undocumented_behaviour_from_v1_9` to false.

Because of the requirement for the configuration option the dependency
for mocha has been updated from `~> 1.3` to `~> 1.11` which shouldn't
have any affect on any projects that depend on this gem.