Fix Minitest 6 deprecations
In Minitest 6 use of global must_equal, must_raise, etc. will fail and
need to be wrapped using `_` or one of its aliases `value` or `expect`.

Rather than doing that this commit rewrites them to use assertions like
`assert_equal`, etc. which follows the style of the rest of the tests.