Remove dependency on rack-cache
This gem is using in a large number of apps and it's always a good idea
to reduce dependencies where possible since each gem's lib directory is
added to the load path and will be scanned when using `require`.

Using rack-cache to parse the Cache-Control header made sense when the
gem was included as a dependency of the actionpack gem but since the
release of Rails 4.0 it's no longer included by default. This means that
the rack-cache dependency is forced on all apps that use this gem,
regardless of whether they use it elsewhere (in fact some apps
explicitly set `config.rack_cache = false`). By reimplementing this
internally we can remove the dependency for all of those apps that
don't need it.

Since the `cache_control` method on the response object is public this
commit reimplements the exact same functionality even though the
response object only ever uses the `max_age` method.