Share a uri_encode helper from the base class
There are two different approaches used in this gem for uri-encoding and
both of them seemed to have problems: the addressable gem was not
required, so asset manager likely only worked in projects with
addressable as a dependency of another gem; Imminence used Erb which is
not a dependency of this project so likely relied on the client app
having erb.

This makes these approaches consistent and sets the groundwork for
providing this functionality in other apps.

I used a `to_s` method since these variables are normally interpolated
within a string (thus an implicit `to_s` call) which would now raise
exceptions for integers. It also seemed that given URI's are a string
based communication mechanism, you'd always need to_s to be called at
some point for usage.