URI encode strings passed into Email Alert API URLs
We see errors in Sentry [1] where clients are requesting information from
Email Alert Frontend with dubious strings that are then passed onto
email Alert API raw. To resolve this I've applied the `uri_encode`
function to all of the parameters that are used in Email Alert API URIs,
this will filter out any characters that should not be in a URL
including hacking attempts. The alternative to this would be to filter
characters in Email Alert Frontend before sending the data to Email
Alert API, this was rejected as it meant duplicating Email Alert API
logic in Email Alert Frontend.

In applying this I decided that this should be done on all URIs for
Email Alert API and not just particular exposed ones, this is because
this repo should not have any knowledge over how these endpoints are
used. Thus, if we consider one exposed to user input we should consider
them all the same.

Ideally this uri encoding would be done implicitly rather than needing
explicit definitions for each uri. It seemed a rather bold thing to
define at this point in the project but something like
`put_json("unsubscribe/:id", params: { id: "test" })` would be preferable.

A pain point revealed by the explicit need was deciding what level of
testing to do. Testing every method felt a bit over-the-top, but that's
the approach I ended up taking since there I felt that it should be
tested and no method should seem special.

[1]: https://sentry.io/organizations/govuk/issues/468258104/?project=202221&query=&statsPeriod=14d