Fix rubocop-govuk 3.17.0 issues
The use of DateTime objects are forbidden in rubocop-govuk now, to
compensate for this I've replaced `Time.now.to_datetime.rfc3339` with
`Time.now.iso8601` which returns the same output (RFC3339 is a subset of
ISO 8601).

A little quirk in Ruby is that the `#iso8601` method is in an extension
of Time provided in the standard library (TIme itself is defined in Ruby
Core) which needs to be required manually. I've added this in so this
code explicitly requires it however in my tests it looks like one of the
dependencies already requires it implicitly.