Fix email-alert-api subscription helpers returning incorrect field
email-alert-api returns an "id" field (see
SubscriptionsController#create), but the test helper returns a
"subscription_id" field.

We're using this field in the account-manager so we can update
someone's email subscription if they update their transition checker
answers.

It's unlikely this change will break anyone's tests as not even
email-alert-frontend uses this field.