Add helper for creating lists in Email Alert API
Currently we stub the adapter in Email Alert Frontend [1], so we're
vulnerable to it changing in future. Note that we can't reuse the
"stub_email_alert_api_creates_subscriber_list" test helper for this
(and assert the requet was made) since the attributes passed to the
helper can be specific to the response (although many will also be
in the request), and we've lost our chance to have a more flexible
"(attributes, returned_attributes)" interface for the helper.

[1]: https://github.com/alphagov/email-alert-frontend/blob/d00c3784fe5b1f7f937b593b8a84082849f879d8/spec/models/content_item_subscriber_list_spec.rb#L17