Support param to skip subscribe confirmation email
Relates to: https://github.com/alphagov/email-alert-api/pull/1503

This is required to prevent sending two emails with conflicting
content / links when a user subscribes via a GOV.UK Account, which
sends its own confirmation email. In order to keep the test stubs
simple, I've switched to using "hash_including" for stubs where
the parameter is (for now) not relevant.