Add a specific HTTPBadRequest exception
This seemed to be a curious omission from the other named exception
classes where the various other common 4xx errors seemed to have named
exceptions.

This has been added in due to challenges of working with exceptions when
an app may return a Bad Request response as per [1].

Unfortunately this repo doesn't have any tests for this error handling
set-up so it wasn't consistent to add a test for this.

[1]: https://github.com/alphagov/collections/blob/abfdf508d30d12704df22bba9349ae9559a445af/app/models/mapit_location.rb#L18-L19