Disable instances of Style/OptionalBooleanParameter
These can't be fixed without causing a backwards compatibility issue.
I've chose to disable these rules with inline comments as we want this
cop to apply for new code.

It is likely that the performance platform aspect of this code will be
removed soon (the platform is due for retirement) which will just leave
the one offence - if I remember I'll see if it can be removed on the
next major release.