Change ruby version to 2.6
This updates the required ruby version of this gem to be 2.6 and sets
the .ruby-version to match. This has been done to reflect that this gem
is only tested with 2.6 and 2.7 [1].

To avoid complications in development and linting the same minor Ruby
version has been used for both, hence the regression from 2.7 to 2.6 in
the .ruby-version file.

[1]: https://github.com/alphagov/govuk-jenkinslib/blob/d5deab4438d803954ef13f94a2cfa528290932d2/vars/govuk.groovy#L617