Remove special case for master branch
This ensures that the `master` branch is published to the Pact broker as
`branch-master` as with all the other branches for consistency. This
also avoids us needing to have this sort of logic in our other apps.

I believe this will also fix an issue where the `master` branch of GDS
API Adapters has been failing since
https://github.com/alphagov/gds-api-adapters/pull/1049 but rather than
re-introduce the logic, I'd prefer `master` to not be treated special.

I notice this was originally introduced in
d12aeb53e4cbd53d6060266358fa859d3cb8b902 but I don't fully understand
the reasoning.