DRY up the Pact tests
Throughout our Pact tests, we've duplicated our intentions, both
in the `will_respond_with` declaration AND in manual RSpec
assertions afterwards (e.g. `assert_equal(reponse_hash, content_id)`).

We should let the Pact gem do the heavy lifting here, and declare
our expected behaviour in just one place. It is up to Pact to fail
the build if the response is not as we expected.

The result should be easier-to-read tests, as we've removed some
nesting in cases where we had `before` blocks which served just
one `it` block. We made a decision in Content Publisher to move
setup code as close as possible to the assertions, which this
commit is enabling: https://docs.publishing.service.gov.uk/apps/content-publisher/testing-strategy.html

I had hoped to remove the 404/409/422 assert_raises statements,
as this is duplication, but the build fails without them as
otherwise they are treated as unexpected exceptions.