Test api/attributes/names endpoint in account-api
This endpoint allows querying if the user has a value stored for an
attribute, without returning the value.

It is very similar to /api/attributes, with the difference that it only
returns the name of the attribute instead of the name and the value.

See where this endpoint was introduced: https://github.com/alphagov/account-api/pull/58

Trello card: https://trello.com/c/e7JQlXx9/710-add-an-endpoint-to-account-api-to-check-if-an-attribute-is-defined