Store response body in new `http_body` field of HTTP exceptions
We don't currently make the response body easily available - it can
only be extracted from the `message` field, by splitting it at the
"Response body:\n" bit, which is a bit unpleasant.  Better for it to
be directly available without manipulating strings.

This is useful for the accounts work, as the account-api is going to
start returning 403 errors when a user doesn't have a sufficiently
high level-of-authentication to access an attribute, and it'll return
the required level-of-authentication in the response body.

An alternative implementation would be to return a 200 response with a
body like:

    { "error": "level of authentication too low", ... }

But this is (1) error prone (you need to remember to check for an
"error" field), and (2) confusing (an error response should have an
error code).  Returning a 403 response, which throws an exception from
gds-api-adapters and so is *guaranteed* to interrupt the normal flow
of control, is much better.

This is a non-breaking change, as the new parameter is the last of
`HTTPErrorResponse` and defaults to `nil`, so anyone constructing
these exceptions in their own code won't have to change things.