First Pact test for Imminence
The general approach of this test is we are testing that imminence returns the correct result when
querying for a `place` based on given coordinates. Hence the coordinates on L20 of test/imminence/imminence_api_pact_test.rb
seem arbitrary, but are deliberately close to but not the same as `place` that is created in the sister Imminence PR
(https://github.com/alphagov/imminence/pull/644), this is to test the proximity logic in Imminence.