Add content_id and title to save-a-page pact tests
I've to do fuzzy matching here, which will check that the response has
fields of the right type.  An alternative would be to make sure the
content IDs and titles match between here and the account-api provider
state, but that then makes it more of a pain to update and I don't
think there's really any advantage.