Permit bearer_token field to be JSON
This enables the bearer_token option for the JsonClient
to be a JSON with the structure '{"value": <string>}'.

The motivation is that we would like to store additional data
with the secret body, to facilitate rotation of the bearer
token purely from the secret body.

E.g. the secret body may look like this:
'{ "api_user": "publisher@alphagov.co.uk", "value": "123",
   "application": "publishing-api" }'

The `value` field would contain the bearer token string.
The remaining fields would enable rotation by a Lambda
with all configuration for the rotation provided in the
body of the secret.