Prepare for sentry-ruby
In govuk_app_config v4, we're swapping out the underlying gem
used to communicate with Sentry, from `sentry-raven` (deprecated)
to the new `sentry-ruby`. There is a [migration guide][] to follow,
but the only thing identified is the old use of `raven_context`,
which should now be `sentry_context`.

I did consider switching to [structured contexts][], which appear
to be best practice, but this doesn't appear to offer any benefits
and could in fact be quite difficult to implement, as we can now
only call `GovukError.configure` once, and we don't want any
explicit references to the Sentry API in gds-api-adapters.
It is lower risk to simply swap out what we had already with the
simplest new alternative.

[migration guide]: https://github.com/getsentry/sentry-ruby/blob/master/MIGRATION.md
[structured contexts]: https://docs.sentry.io/platforms/ruby/enriching-events/context/#structured-context

Trello: https://trello.com/c/xo4l4GWQ/2587-update-gds-api-adapters-to-use-pre-released-govukappconfig-3