Forbid creating a SubscriberList with a content_id AND tags or links
In [1] we changed email-alert-api to allow signing up to notifications
about a specific content ID.  We're now building the user journeys for
this feature, so we'll be creating such subscriber lists.

It doesn't make sense to create a list with both a `content_id`
restriction *and* `tags` or `links`, because the `content_id` already
restricts potential email triggers to that single page: additional
filtering adds nothing, and runs the risk of creating a list which can
never trigger emails.

[1] https://github.com/alphagov/email-alert-api/pull/1640