Remove attribute types from account API
We are simplifying account API and removing the `local` attribute type.

This means it will only store what were previously `cached` attributes
(those provided by the IDP), and therefore aren't allowed to be changed
by any GOV.UK app.

Remove the pact test that tries to set an attribute in preparation for
this.