Don’t require govuk_request_id in Email Alert API bulk_unsubscribe stub
The GOVUK-Request-Id HTTP header is set by Nginx on an initial request and
passed by the apps to the APIs that they call [^1]. The gds-api-adapters handle
the header in the middleware [^2].

It is only necessary to explicitly pass the header for out-of-band processes
via asynchronous workers (e.g. Sidekiq).

This still allows to pass it in as a parameter so it's not a breaking change
and also updates the documentation to more accurately describe when the param
is required.

1. https://github.com/alphagov/govuk-puppet/blob/05af1/modules/nginx/templates/proxy-vhost.conf#L7-L25
2. https://github.com/alphagov/gds-api-adapters/blob/2bb3a/lib/gds_api/railtie.rb#L5-L8