Update gems
This goes through and updates the gem definitions for the development
dependencies. This has been done to avoid having dependabot raise a
bunch of PRs (and as just a bit of a spring clean).

It removes the dependencies on pry, yard and webrick as there are not
references to them.

It also removes the version pinning for Rake since it's a stable gem
that seems to often end up behind major versions in our gems.