Turn off pact analytics tracking
Pact have recently added tracking to their project which will result in
data sent from our build servers to their servers [1].

I'm not sure if we have a collective opinion on whether we're happy with
this or not, so it seems simplest to disable it and save the need for a
divisive decision. It also gets rid of an annoying warning.

[1]: pact-foundation/pact-ruby#256