Generate and publish pact tests
This addition is to replicate the similar functionality in the
Jenkinsfile [1].

This functionality is used to publish pact JSON files to the pact
broker, it is tagged with the branch name.

The Jenkinsfile relied on being able to publish the pact JSON files that
were generated by running the tests. This is harder to achieve in a
GitHub action because each matrix runs on a separate machine.

I felt that it was simpler to understand having this step run separately
from the matrix, but to do that we need to generate the pact JSON files.
I therefore added in an extra rake task that can run just the pact tests
to generate them.

[1]: https://github.com/alphagov/gds-api-adapters/blob/2f003aa7a5b808e4bd7d57224545abd6fc606af8/Jenkinsfile#L89-L93