Configure apps for pact test
This triggers builds on applications that have a pact with this utility.
It does this by using workflow files that are defined in each of the
repositories which can run the job. These are currently not merged into
the repos while I seek feedback on the approach, hence the non-main
branch.

This is a different approach than the Jenkinsfile takes which clones the
applications and runs the commands directly [1]. The reason for this is
that the configuration for these other apps hasn't aged well (for
example they test against the wrong versions of the databases and need
some manual set-up [2]). Our theory is that if these files are with the
app, and become part of regular builds, they stand a much higher chance
of being maintained.

To implement this I did see if it was possible to use a matrix strategy
so they would be less verbose. However this relied on interpolation in
the workflow name, and this is something that GitHub actions currently
don't support.

[1]: https://github.com/alphagov/gds-api-adapters/blob/2f003aa7a5b808e4bd7d57224545abd6fc606af8/Jenkinsfile#L95-L110
[2]: https://github.com/alphagov/gds-api-adapters/commit/fdbce2503de22171c814744433c950ae714181df