Move pact tests to pacts directory
This is to make it easier to identify pact tests from regular tests and
make their organisation consistent. Link Checker API was unusual as it had
a mixture of pact tests and unit tests. I've separated them out.

These tests are quite different from the unit test, they don't
actually assert behaviour - they instead generate JSON files that
applications use to assert behaviour.

I intend to separate out the task to run the tests into separate ones
for pact and unit tests, this is to allow them to be generated at a
different stage in the CI process.