Fix pact-target-branch for PRs
github.ref_name can't be reliably used as a branch name. When the GitHub
Action is running for a pull_request this isn't set to the branch name
and is instead based on the PR number (for example "1176/merge" [1]) and
is instead in the variable github.head_ref. I couldn't find a great way
this was documented unfortunately, the docs [2] are a bit confusing.

I didn't want to have the OR condition multiple times in the code
(particularly as it's rather opaque) so I used a GitHub Action output as
a way to set this to be something more meaningful.

[1]: https://github.com/alphagov/gds-api-adapters/actions/runs/3593090463/jobs/6049617350
[2]: https://docs.github.com/en/actions/learn-github-actions/contexts#github-context