Allow setting a pattern for publishing pacts
I've added this option as I want to enable the ability to upload pacts
that have been previously generated and thus might not be in the regular
pact directory.

My expected use case for this is to create the Pacts, upload them to
GitHub Actions as an artifact, then in a later job download them and
publish them to the pact broker.