Use GitHub Artifacts to distribute pacts to apps for testing
This changes the approach for testing downstream apps against newly
generated pact tests. We previously would upload the pact tests to the
Pact Broker and then have the apps download these from the Pact broker
to test against them. This changes that approach to instead use a GitHub
Action Artifact [1] as a place to store these files for testing.

The motivation for this change is because the Pact Broker method
requires permission to view GitHub Action secrets, without this the task
fail. We feel this pain most noticeably on pull requests opened by
Dependabot.

Applications that test against these pacts (consumers) will now operate in
two ways depending on the circumstance. When changes are made to them or
they are being locally tested they will pull the `branch-main` pacts
from the pact-broker as per before. When they test against an iteration
of GDS API Adpaters they will access the pacts from an artifact.

---

This change separates the generate and publish pact steps into two, now
only publishing the pacts when on the main branch. This will mean that
eventually there is only branch-main on the pact-broker.

All the downstream applications need changes to their pact-verify.yml
action to accommodate this change, hence why they reference
@pact-artifact branches.

I didn't see their being value in keeping the artifact after testing so
I set-up a means to delete it. This is a job that will only run if
generate_pacts was a success and not until publish_pacts has been run or
skipped. It will run even if publish_pacts was a failure.

[1]: https://docs.github.com/en/actions/using-workflows/storing-workflow-data-as-artifacts