Change source for `GdsApi.worldwide.world_locations` method
This method currently returns data retrieved from the Whitehall API,
which is being deprecated.

Therefore updating the method to get the data from the World Content
Item.

This is a breaking change as we have removed the pagination from the
response. This is to avoid multiple requests to the content item, when
all the results are included anyway.