Change data source for `organisations_for_world_location` method
This ends reliance on the Whitehall API by changing the data source to
use Search API and Content Store.

The output is in an identical format, but is no longer paginated.

Also increasing the test coverage by ensuring we are correctly
transforming the data. Previously we only checked that the title was
correct.

We are having to make multiple requests to Content Store (one per
organisation found in Search API) as there is no current way of
retreiving this data in bulk. We believe very few organisations have
more than a handful of organisations, so the number of requests will be
limited.